% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DECODE-package.R
\docType{package}
\name{DECODE-package}
\alias{DECODE-package}
\title{DECODE: Deciphering Cancer Origin from DNA Evolution}
\description{
Decomposes genomic variants from bulk DNA-sequencing into a neutral tail
and mutation clusters based on their Variant Allele Frequencies (VAFs), with
corrections for sample-specific DNA-sequencing coverage distribution and
mutation calling biases in the Site Frequency Spectrum (SFS).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dinhngockhanh/DECODE}
  \item Report bugs at \url{https://github.com/dinhngockhanh/DECODE/issues}
}

}
\author{
\strong{Maintainer}: Khanh N. Dinh \email{dinhngockhanh@gmail.com}

}
\keyword{internal}
