% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DECODE_analysis.r
\name{DECODE_mutation_assignment}
\alias{DECODE_mutation_assignment}
\title{Assign mutations to DECODE SFS components}
\usage{
DECODE_mutation_assignment(
  DECODE_result,
  mutation_table,
  mode = "inference_A",
  with_tail = NA,
  N_clusters = NULL
)
}
\arguments{
\item{DECODE_result}{Object returned by function \code{\link{DECODE}}.}

\item{mutation_table}{Mutational dataframe.
Each row corresponds to a mutation, which can be associated with different copy number states
(unlike the mutation table for \code{\link{DECODE}}, which should include only mutations with the same copy number background).
\code{mutation_table} must contain column \code{normalized_VAF},
consisting of VAFs normalized to match the copy number state analyzed by \code{\link{DECODE}}.}

\item{mode}{String to specify which DECODE mode to use:
\code{"inference_A"}, \code{"inference_B"}, or \code{"validation"} (\code{"inference_A"} by default).}

\item{with_tail}{Logical variable for whether to apply DECODE fit with or without tail:
\code{TRUE}, \code{FALSE}, or \code{NA} (\code{NA} by default).
If \code{with_tail = NA}, the best overall fit chosen by \code{\link{DECODE}} is used.}

\item{N_clusters}{Integer to select the cluster count in the DECODE fit to apply (\code{NULL} by default).
If \code{N_clusters = NULL}, the best overall fit chosen by \code{\link{DECODE}} is used.}
}
\value{
Dataframe extending \code{mutation_table} with probability columns:
\code{prob_tail} = probability that each mutation belongs to the tail (if a DECODE fit with tail is selected).
\code{prob_cluster_*} = probability that each mutation belongs to cluster \code{*}.
}
\description{
Infer the probability for each mutation to belong to a SFS component
as inferred by DECODE.
}
