% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getAssoGeneSetPValue.r
\name{getAssoGeneSetPValue}
\alias{getAssoGeneSetPValue}
\title{Calculate the p-value between selected genes and functional gene set}
\usage{
getAssoGeneSetPValue(geneList, geneSet, multipleTestCount, MaxGene)
}
\arguments{
\item{geneList}{Selected genes}

\item{geneSet}{Functional gene set}

\item{multipleTestCount}{Number of multiple testing}

\item{MaxGene}{Number of genes in expression data}
}
\value{
The adjusted p-value for the associated gene set
}
\description{
Calculate the p-value between selected genes and functional gene set
}

