% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getBestAssociatedGeneSet.r
\name{getBestAssociatedGeneSet}
\alias{getBestAssociatedGeneSet}
\title{Get best associated functional gene sets for partitions of gene i}
\usage{
getBestAssociatedGeneSet(pathway, all8Partitions, onePartition, MaxGene,
  minSupport)
}
\arguments{
\item{pathway}{All functional gene sets}

\item{all8Partitions}{All eight possible partitions for gene i}

\item{onePartition}{The partition to be associated with the functional gene set}

\item{MaxGene}{Number of genes in expression data}

\item{minSupport}{Minimum support for functional gene set}
}
\value{
The adjusted p-values for the best associated gene set of the input partition
}
\description{
Get best associated functional gene sets for partitions of gene i
}

