% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getBonferroniPValue.r
\name{getBonferroniPValue}
\alias{getBonferroniPValue}
\title{Adjust p-value by Bonferroni correction}
\usage{
getBonferroniPValue(pValues)
}
\arguments{
\item{pValues}{Unadjusted p-values}
}
\value{
Adjusted p-values
}
\description{
Adjust p-value by Bonferroni correction
}

