% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getDE_DC_OptimalThreshold.r
\name{getDE_DC_OptimalThreshold}
\alias{getDE_DC_OptimalThreshold}
\title{Perform chi-square optimization}
\usage{
getDE_DC_OptimalThreshold(t_result, MaxGene, d_r, minSupport)
}
\arguments{
\item{t_result}{The t-statistics}

\item{MaxGene}{Number of genes in expression data}

\item{d_r}{DC measures}

\item{minSupport}{The minimum expected frequency in contingency table}
}
\value{
The optimal threshold information
}
\description{
Perform chi-square optimization
}

