% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getFDR.r
\name{getFDR}
\alias{getFDR}
\title{Adjust p-value by Benjamini and Hochberg method}
\usage{
getFDR(pValues)
}
\arguments{
\item{pValues}{Unadjusted p-values}
}
\value{
Adjusted p-values
}
\description{
Adjust p-value by Benjamini and Hochberg method
}

