% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getPartitionIndex.r
\name{getPartitionIndex}
\alias{getPartitionIndex}
\title{Get gene index of 8 partitions for gene i}
\usage{
getPartitionIndex(gene_i, t_result, optimalCutOff, abs_r)
}
\arguments{
\item{gene_i}{Gene i index}

\item{t_result}{t-statistics}

\item{optimalCutOff}{Optimal thresholds}

\item{abs_r}{Matrix consisting of absolute values of all differential co-expression measures}
}
\value{
The selected genes for each partition in index
}
\description{
Get gene index of 8 partitions for gene i
}

