% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getPathway.r
\name{getPathway}
\alias{getPathway}
\title{read functional gene sets}
\usage{
getPathway(inputFile, geneName, minSupport)
}
\arguments{
\item{inputFile}{Input file name}

\item{geneName}{Gene name lists}

\item{minSupport}{Minimum support}
}
\value{
Functional gene set
}
\description{
read functional gene sets
}

