% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/openFileToWrite.r
\name{openFileToWrite}
\alias{openFileToWrite}
\title{Open file to write result}
\usage{
openFileToWrite(filename)
}
\arguments{
\item{filename}{file name
Output: Results in text file}
}
\description{
Open file to write result
}

