% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sumResult_MinGain.r
\name{sumResult_MinGain}
\alias{sumResult_MinGain}
\title{Summarize the functional gene set results into text file}
\usage{
sumResult_MinGain()
}
\description{
Summarize the functional gene set results into text file
}

